Xilsecure:Use of mixed mode arithmetic.
This patch fixes below MISRA-C violation:
Use of mixed mode arithmetic.
MISRA-C:2012/AMD1 R.10.1,R.10.2,R.10.3,R.10.4


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilsecure/src/common/xsecure_rsa.c
@@ -317,7 +317,7 @@ s32 XSecure_RsaPublicEncrypt(XSecure_Rsa *InstancePtr, u8 *Input, u32 Size,
 s32 XSecure_RsaPrivateDecrypt(XSecure_Rsa *InstancePtr, u8 *Input, u32 Size,
 				u8 *Result)
 {
-	s32 Status = XSECURE_RSA_DATA_VALUE_ERROR;
+	s32 Status = (s32)XSECURE_RSA_DATA_VALUE_ERROR;
 	u32 idx;
 
 	/* Assert validates the input arguments */
